ai/09-roboflow-supervison/football/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiohttp-retry==2.8.3
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.8.0
APScheduler==3.11.0
asttokens==3.0.0
async-timeout==5.0.1
asyncua==1.1.5
attrs==24.3.0
av==12.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
boto3==1.35.60
botocore==1.35.99
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.110.3
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gdown==5.2.0
google-crc32c==1.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.25.2
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.7
ifaddr==0.2.0
imageio==2.36.1
inference==0.34.0
inference-gpu==0.29.2
iniconfig==2.0.0
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.19.2
onnxruntime-gpu==1.18.0
openai==1.59.8
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
paho-mqtt==1.6.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pluggy==1.5.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pybase64==1.0.2
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydot==2.0.0
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.10.0
pylogix==1.0.5
pynndescent==0.5.13
pyOpenSSL==25.0.0
pyparsing==3.2.1
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.0.1
roboflow==1.1.50
s3transfer==0.10.4
safetensors==0.5.2
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
slack_sdk==3.33.5
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sports @ git+https://github.com/roboflow/sports.git@06053616f1f8a8ae1fa936eb00dcdc2e4f888bb1
stack-data==0.6.3
starlette==0.37.2
structlog==24.4.0
supervision==0.25.1
sympy==1.13.1
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2025.1.10
tldextract==5.1.3
tokenizers==0.20.3
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
triton==3.1.0
twilio==9.3.8
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ultralytics==8.0.196
ultralytics-thop==2.0.13
umap-learn==0.5.7
urllib3==2.3.0
wcwidth==0.2.13
yarl==1.18.3
zxing-cpp==2.2.0
